changed tower upgrade cost rate and background images additions
1. Increased the tower upgrade cost eponent base from 1.3 to 1.5 so it would round up better.

2. created 3 linear and 3 circular types of backgrounds, one for each type of base color(RGB).

3. Made the circular types take a center position so it is allways centered on the Base.

+++ Maps.py
@@ -34,7 +34,7 @@ class map:
 
 class TestMap(map):
     map_pixels = [(0, 50), (1200, 200), (1200, 700), (200, 700), (200, 400), (700, 450)]
-    base_position = (screenWidth * widthMultiplier / 2, screenHeight * heightMultiplier / 2)
+    base_position = (int(screenWidth * widthMultiplier / 2), int(screenHeight * heightMultiplier / 2))
     map.convert(map,map_pixels)
     
 class LoopyMap(map):
